add email and username to User objects
