--- src/main/java/se/myhappyplants/api/user/NewUserRequest.java
@@ -1,14 +1,0 @@
-package se.myhappyplants.api.user;
-
-public class NewUserRequest {
-    public String name;
-    public String email;
-
-    public NewUserRequest() {
-    }
-
-    public NewUserRequest(String name, String email) {
-        this.name = name;
-        this.email = email;
-    }
-}
