--- src/main/java/se/myhappyplants/api/user/User.java
@@ -1,13 +1,0 @@
-package se.myhappyplants.api.user;
-
-public class User {
-    public int id;
-    public String name;
-    public String email;
-
-    public User(int id, String name, String email) {
-        this.id = id;
-        this.name = name;
-        this.email = email;
-    }
-}
